AoC 2022 Day7

+++ src/aoc2022/day_11/ItemThrow.kt
@@ -1,6 +1,6 @@
 package aoc2022.day_11
 
 data class ItemThrow(
-    val item: Int,
+    val item: Long,
     val targetIndex: Int
 )
